requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b3
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.16.0
azure-search-documents==11.4.0
bidict==0.23.1
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.505
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.5.14
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.110.3
fastapi-socketio==0.0.10
filetype==1.2.0
frozenlist==1.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.0
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.51
Lazify==0.4.0
literalai==0.0.507
marshmallow==3.21.1
msal==1.28.0
msal-extensions==1.1.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.23.6
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.1
packaging==23.2
portalocker==2.8.2
protobuf==4.25.3
pycparser==2.22
pydantic==2.7.1
pydantic-core==2.18.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.6.0
tomli==2.0.1
tqdm==4.66.2
types-requests==2.31.0.20240406
typing-extensions==4.11.0
typing-inspect==0.9.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
